feat: add `origin` config parameter
This commit adds the `origin` as an optional configuration parameter
that can be set in the same way as `host` and `port`.

This closes https://github.com/tuono-labs/tuono/issues/232 and solves
use-cases such as being able to develop on localhost with HTTPS with an
origin name such as `https://tuono.localhost`, where HTTPS is required
for testing certain web features (relating to e.g. storage or web
workers).

Prior to this commit this will not work as the Vite tools will attempt
to load resources from e.g. `http://localhost:3000` and be blocked by
the browser for attempting to load HTTP resources from an HTTPS web
page.